+++ src/main/java/com/beimin/eveapi/response/eve/StationListResponse.java
@@ -1,19 +1,19 @@
 package com.beimin.eveapi.response.eve;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import com.beimin.eveapi.model.eve.Station;
 import com.beimin.eveapi.response.ApiResponse;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class StationListResponse extends ApiResponse {
-	private final Map<Integer, Station> stations = new HashMap<Integer, Station>();
+	private final Map<Long, Station> stations = new HashMap<Long, Station>();
 
 	public void add(Station station) {
 		stations.put(station.getStationID(), station);
 	}
 
-	public Map<Integer, Station> getStations() {
+	public Map<Long, Station> getStations() {
 		return stations;
 	}
 }
